Fix CRM_Contact_Import_Parser_Contact::createContact() to not throw an unhandled exception that stops the Import and fails to report which imported row failed. This occurs when the matching is done on an External ID and names+email is not supplied and the supplied External ID is not found. Note that in this case $this->_updateWithId is true.
Catch the exception and set $newContact to trigger the "Not a dupe, so we
had an error" code at the end of CRM_Contact_Import_Parser_Contact::import.

Moved the code that handles $this->_updateWithId at the end of
CRM_Contact_Import_Parser_Contact::import() so that the "Not a dupe, so we
had an error" code has a chance to be invoked. This seems reasonable as the
$this->_updateWithId shouldn't be called if there has been a
CRM_Import_Parser::ERROR. Without this change, testing shows the error is
reported but the CSV of errors is misformed.