Add support for enrolling people in Thinkific bundles (note that cbf_course_thinkific_enrolment() already supports this).
To do this we need to know whether $course is the Id of a Thinkific course
or a Thinkific bundle. That is stored in the CiviCRM OptionValue which
corresponds to this $course, stored in the 'description' as the word
'bundle'.

If that OptionValue can't be found, the enrolment fails as we have a
configuration issue.